requirements.txt
torch>=1.7.0
numpy>=1.19.0
numba>=0.50.0
ultralytics>=8.0.0
Pillow>=8.0.0
torchvision>=0.8.0 